// src/daodao/types/contracts/DaoVotingCw721Staked.ts
/**
 * This file was automatically generated by @cosmwasm/ts-codegen@1.10.0.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @cosmwasm/ts-codegen generate command to regenerate this file.
 */

export type ActiveThreshold =
  | {
      absolute_count: {
        count: Uint128
      }
    }
  | {
      percentage: {
        percent: Decimal
      }
    }
export type Uint128 = string
export type Decimal = string
export type NftContract =
  | {
      existing: {
        address: string
      }
    }
  | {
      new: {
        code_id: number
        initial_nfts: Binary[]
        label: string
        msg: Binary
      }
    }
  | {
      factory: Binary
    }
export type Binary = string
export type Duration =
  | {
      height: number
    }
  | {
      time: number
    }
export interface InstantiateMsg {
  active_threshold?: ActiveThreshold | null
  nft_contract: NftContract
  unstaking_duration?: Duration | null
}
export type ExecuteMsg =
  | {
      receive_nft: Cw721ReceiveMsg
    }
  | {
      unstake: {
        token_ids: string[]
      }
    }
  | {
      claim_nfts: {
        // v2.6.0 and above.
        type?: ClaimType
      }
    }
  | {
      update_config: {
        duration?: Duration | null
      }
    }
  | {
      add_hook: {
        addr: string
      }
    }
  | {
      remove_hook: {
        addr: string
      }
    }
  | {
      update_active_threshold: {
        new_threshold?: ActiveThreshold | null
      }
    }
export type ClaimType =
  | 'legacy'
  | 'all'
  | {
      specific: string[]
    }
export interface Cw721ReceiveMsg {
  msg: Binary
  sender: string
  token_id: string
}
export type QueryMsg =
  | {
      config: {}
    }
  | {
      nft_claims: {
        address: string
      }
    }
  | {
      hooks: {}
    }
  | {
      staked_nfts: {
        address: string
        limit?: number | null
        start_after?: string | null
      }
    }
  | {
      active_threshold: {}
    }
  | {
      is_active: {}
    }
  | {
      voting_power_at_height: {
        address: string
        height?: number | null
      }
    }
  | {
      total_power_at_height: {
        height?: number | null
      }
    }
  | {
      dao: {}
    }
  | {
      info: {}
    }
export interface ActiveThresholdResponse {
  active_threshold?: ActiveThreshold | null
}
export type Addr = string
export interface Config {
  nft_address: Addr
  unstaking_duration?: Duration | null
}
export interface HooksResponse {
  hooks: string[]
}
export interface InfoResponse {
  info: ContractVersion
}
export interface ContractVersion {
  contract: string
  version: string
}
export type Boolean = boolean
export type Expiration =
  | {
      at_height: number
    }
  | {
      at_time: Timestamp
    }
  | {
      never: {}
    }
export type Timestamp = Uint64
export type Uint64 = string
export interface NftClaimsResponse {
  nft_claims: NftClaim[]
}
export interface NftClaim {
  release_at: Expiration
  token_id: string
  // v2.6.0 and above.
  legacy?: boolean
}
export type ArrayOfString = string[]
export interface TotalPowerAtHeightResponse {
  height: number
  power: Uint128
}
export interface VotingPowerAtHeightResponse {
  height: number
  power: Uint128
}